Clean up hack-session code structure
* Rename mail.js -> hack-session.js
* Create folder named hack-session and move ip.js there
* Split out mail specific functionality and create mail.js in the folder
